Rename common test file
