+++ tests/puppeteer/campaigns/functional/BO/catalog/brandsAndSuppliers/brands/04_BulkActionsBrands.js
@@ -4,7 +4,7 @@ const helper = require('@utils/helpers');
 const files = require('@utils/files');
 const BrandFaker = require('@data/faker/brand');
 const loginCommon = require('@commonTests/loginBO');
-const brandsCommon = require('@commonTests/BO/brands');
+const brandsCommon = require('@commonTests/BO/BOBase');
 // Importing pages
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
