+++ tests/puppeteer/campaigns/functional/BO/catalog/brandsAndSuppliers/brands/05_BulkActionsBrandAddresses.js
@@ -3,7 +3,7 @@ const {expect} = require('chai');
 const helper = require('@utils/helpers');
 const BrandAddressFaker = require('@data/faker/brandAddress');
 const loginCommon = require('@commonTests/loginBO');
-const brandsCommon = require('@commonTests/BO/brands');
+const brandsCommon = require('@commonTests/BO/BOBase');
 // Importing pages
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
